% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.ebmr}
\alias{coef.ebmr}
\title{extract regression coefficients from ebmr fit}
\usage{
\method{coef}{ebmr}(object, ...)
}
\arguments{
\item{object}{an ebmr fit}
}
\value{
a p vector of estimated regression coefficients
}
\description{
extract regression coefficients from ebmr fit
}
