% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebmr.update.grr.R
\name{ebmr.update.grr.svd}
\alias{ebmr.update.grr.svd}
\title{Update by fitting GRR model}
\usage{
ebmr.update.grr.svd(
  fit,
  tol = 1e-08,
  maxiter = 1000,
  compute_Sigma_diag = TRUE,
  compute_Sigma_full = FALSE,
  update_residual_variance = TRUE
)
}
\arguments{
\item{fit}{a previous ebmr fit}

\item{tol}{a tolerance for the EM algorithm}

\item{maxiter}{maximum number of iterations for EM algorithm}

\item{compute_Sigma_diag}{boolean flag; set to TRUE to compute the diagonal of the posterior variance, otherwise this variance is set to 0.}

\item{compute_Sigma_full}{boolean flag; set to TRUE to compute full Sigma matrix for testing purposes}

\item{update_residual_variance}{boolean flag; set to FALSE to keep residual variance fixed}
}
\value{
an updated ebmr fit
}
\description{
Updates parameters of an EBMR fit by fitting the full GRR model,
including estimating hyperparameters (residual variance and prior scaling factor).
The method uses an SVD on Xtilde = XW^0.5, after which
other computations (EM algorithm to maximize hyperparameters, and posterior computations)
are cheap.
}
\details{
The code is based on the ideas at https://stephens999.github.io/misc/ridge_em_svd.html
but extended to deal with case where n>p as well as n<=p.
}
