% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnv.R
\name{ebnv.np}
\alias{ebnv.np}
\alias{ebnv.exp}
\alias{ebnv.pm}
\alias{ebnv.exp_mix}
\title{Solve EBNV problem}
\usage{
ebnv.np(b, s2, g)

ebnv.exp(b, s2, g = NULL)

ebnv.pm(b, s2, g = NULL)

ebnv.exp_mix(b, s2, g)
}
\arguments{
\item{b}{vector of n observations}

\item{s2}{a positive real number}

\item{g}{(used only in some cases) an object with the same structure as the prior to be fit
used to specify things like the grid used for discrete priors.}
}
\value{
A list with elements

\item{g}{A list containing the details of the estimated prior}

\item{wbar}{A vector containing inverses of posterior mean for 1/w}
}
\description{
Functions for solving the Empirical Bayes Normal Variances (EBNV) problem for various
prior families
}
\details{
These functions fit the model

b_j~N(0,s2 w_j)

w_j ~ g()

by empirical
Bayes. They estimate the prior g, and (the inverse of the) posterior mean for each 1/w_j.
}
\section{Functions}{
\itemize{
\item \code{ebnv.np}: Solve EBNV problem with non-parametric prior

\item \code{ebnv.exp}: Solve EBNV problem with exponential prior

\item \code{ebnv.pm}: Solve EBNV problem with point mass prior

\item \code{ebnv.exp_mix}: Solve EBNV problem with mixture of exponentials prior
}}

