% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnv.R
\name{ebnv_np}
\alias{ebnv_np}
\title{Solve EBNV problem with non-parametric prior}
\usage{
ebnv_np(b, s2, g)
}
\arguments{
\item{b}{n-vector of observations}

\item{s2}{a scalar variance parameter}

\item{g}{a list with elements mixprop and w, the latter being a non-negative grid of possible values (unlike ash, 0 not allowed)}
}
\value{
a list with elements g and wbar (inverse of posterior mean for 1/w)
}
\description{
Solves the EBNV problem for non-parametric prior (discrete grid)
}
\details{
fits the model b_j \sim N(0,s2 w_j) and w_j \sim g() by empirical
Bayes, where g is non-parametric distribution approximated by a discrete distribution
on a grid
}
