% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebmr.init.R
\name{ebmr.init}
\alias{ebmr.init}
\title{Initialize ebmr fit}
\usage{
ebmr.init(X, y, sb2 = 1)
}
\arguments{
\item{X}{a numeric n times p matrix}

\item{y}{an n vector}

\item{sb2}{scalar parameter value}
}
\description{
Initialize ebmr fit
}
