% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebmr.R
\name{ebmr}
\alias{ebmr}
\title{Fit Empirical Bayes Multiple Regression Model}
\usage{
ebmr(X, y, tol = 1e-10, maxiter = 1000, ebnv_fn = ebnv_exp)
}
\arguments{
\item{X}{An n times p numeric matrix of covariates.}

\item{y}{An n vector of responses.}

\item{tol}{Small real number controlling convergence tolerance;
algorithm stops when elbo changes less than \code{tol}.}

\item{maxiter}{Integer indicating maximum number of iterations.}

\item{admm}{If \code{admm = TRUE}, use the ADMM updates for
updating the regression coefficients.}

\item{k}{Integer indicating the rank of the approximation to use
for Sigma update (default \code{NULL} uses full rank)}

\item{thin}{Integer indicating how often to update Sigma (e.g.,
\code{thin = 10} means update every 10 iterations).}
}
\value{
An object of class "ebmr" that contains fit details.
}
\description{
This is a new approach based on generalized ridge
  regression.
}
