% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnv.R
\name{ebnv.np}
\alias{ebnv.np}
\alias{ebnv.exp}
\alias{ebnv.pm}
\alias{ebnv.exp_mix}
\title{Solve EBNV problem}
\usage{
ebnv.np(b, s2, g.init)

ebnv.exp(b, s2)

ebnv.pm(b, s2)

ebnv.exp_mix(
  b,
  s2,
  g.init,
  update.mixprop = c("em", "mixsqp", "none"),
  update.w = c("em", "none")
)
}
\arguments{
\item{b}{vector of n observations}

\item{s2}{a positive real number}

\item{g.init}{(present only in some cases) an object with the same structure as the prior to be fit
used to initialize the algorithm (and so, implicitly, to sometimes specify parameters like the grid to be used)}

\item{update.mixprop}{string indicating how to estimate/update the mixture proportions; if "none" then mixture proportions are supplied by g$mixprop}

\item{update.w}{string indicating how to update w parameters; ; if "none" then mixture proportions are supplied by g$w}
}
\value{
A list with elements

\item{g}{A list containing the details of the estimated prior}

\item{wbar}{A vector containing inverses of posterior mean for 1/w}

\item{loglik}{The log-likelihood p(b | s2, ghat)}
}
\description{
Functions for solving the Empirical Bayes Normal Variances (EBNV) problem for various
prior families
}
\details{
These functions fit the model

b_j~N(0,s2 w_j)

w_j ~ g()

by empirical
Bayes. They estimate the prior g, and (the inverse of the) posterior mean for each 1/w_j.
}
\section{Functions}{
\itemize{
\item \code{ebnv.np}: Solve EBNV problem with non-parametric prior

\item \code{ebnv.exp}: Solve EBNV problem with exponential prior

\item \code{ebnv.pm}: Solve EBNV problem with point mass prior

\item \code{ebnv.exp_mix}: Solve EBNV problem with mixture of exponentials prior
}}

