% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnv.R
\name{ebnv_np}
\alias{ebnv_np}
\title{Solve EBNV problem with non-parametric prior}
\usage{
ebnv_np(b, s2, wgrid)
}
\arguments{
\item{b}{n-vector of observations}

\item{s2}{a scalar variance parameter}

\item{wgrid}{a discrete grid of non-negative values (unlike ash, 0 not allowed)}
}
\value{
a list with elements mixprop, wgrid and wbar (inverse of posterior mean for 1/w)
}
\description{
Solves the EBNV problem for non-parametric prior (discrete grid)
}
\details{
fits the model b_j \sim N(0,s2 w_j) and w_j \sim g() by empirical
Bayes, where g is non-parametric distribution approximated by a discrete distribution
on a grid
}
